Refactors to channel model and move tests to new branch. (#102)
Make the following changes:
1. Remove the "Multiselect" fallthrough and related functions and just
use separate functions for each number of cases
2. Make it so we only select randomly on the first attempt. 
3. Add an explicit return statement to the end of Send
4. Added the channel tests I had in the master branch to the new branch

It was getting to be a real pain to try to use a shared 5 case
fallthrough for the different numbers of cases

Using a single random attempt instead of a permutation is easier to
prove and technically still ensures every immediately selectable case
could be selected. In fact, it really isn't any different fairness wise
than before if we don't create a new permutation on every loop.

We don't currently support void functions without explicit return
statements if they happen to have a call to a non-void function on the
last line. I filed an issue for this